- fix(adc): resolve channel data corruption after hardware calibration - feat(sysctrl): update implementation of SYSCTRL_EnableDCDCMode - refactor(example/voice_remote_ctrl): unify ADC and ASDM into analog module
