+++ examples/voice_remote_ctrl/voice_remote_ctrl.ingw
@@ -3,11 +3,11 @@
   "var": {
    "link_layer_features": {
     "value": "bf79ff07fe000000",
-    "addr": 33718312
+    "addr": 33718300
    }
   },
   "rom": {
-   "size": 164848,
+   "size": 164836,
    "base": 33562624
   },
   "ram": {
@@ -18,11 +18,11 @@
    "base": 33730560
   },
   "version": [
-   2,4,20]
+   2,4,25]
  },
- "tc_dir": "C:\\Keil_v5\\UV4",
+ "tc_dir": "C:\/Keil_v5\/uv4",
  "ide_ver": "5",
- "ide_exe": "C:\\Keil_v5\\UV4\\UV4.exe",
+ "ide_exe": "C:\/Keil_v5\/uv4\/uv4.exe",
  "ota": false,
  "ide": "keil",
  "git": false,
@@ -41,7 +41,7 @@
  "pb_adv": true,
  "pb_gatt": true,
  "deep_sleep": true,
- "proj_dir": "C:\/projects\/ingchips\/SDK_Release\/examples\\voice_remote_ctrl",
+ "proj_dir": "C:\/ming\/source\/ING918XX_SDK_SOURCE\/examples\\voice_remote_ctrl",
  "power_saving": false,
  "gatt_handle_value": "6",
  "bundle": "typical",
