+++ examples/voice_remote_ctrl/voice_remote_ctrl.uvproj
@@ -11,9 +11,9 @@
       <uAC6>0</uAC6>
       <TargetOption>
         <TargetCommonOption>
-          <Device>ing91800</Device>
+          <Device>ing91600</Device>
           <Vendor>INGCHIPS</Vendor>
-          <Cpu>CLOCK(32000000) CPUTYPE("Cortex-M3") ELITTLE IROM(0x4000-0x43fff) IRAM(0x20000000-0x2000ffff)</Cpu>
+          <Cpu>CLOCK(128000000) CPUTYPE("Cortex-M4") FPU2 ELITTLE IROM(0x2000000-0x21fffff) IRAM(0x20000000-0x20007fff)</Cpu>
           <FlashUtilSpec/>
           <StartupFile/>
           <FlashDriverDll/>
@@ -312,8 +312,8 @@
               </OCR_RVCT3>
               <OCR_RVCT4>
                 <Type>1</Type>
-                <StartAddress>163840</StartAddress>
-                <Size>376832</Size>
+                <StartAddress>33730560</StartAddress>
+                <Size>348160</Size>
               </OCR_RVCT4>
               <OCR_RVCT5>
                 <Type>1</Type>
@@ -337,8 +337,8 @@
               </OCR_RVCT8>
               <OCR_RVCT9>
                 <Type>0</Type>
-                <StartAddress>536900888</StartAddress>
-                <Size>35560</Size>
+                <StartAddress>536892600</StartAddress>
+                <Size>11080</Size>
               </OCR_RVCT9>
               <OCR_RVCT10>
                 <Type>0</Type>
@@ -376,7 +376,7 @@
               <MiscControls>--gnu</MiscControls>
               <Define>USE_STDPERIPH_DRIVER,AUDIO_INPUT=0,AUDIO_CODEC_ALG=0</Define>
               <Undefine/>
-              <IncludePath>.\libsbc\include;../../bundles/typical/inc;../../bundles/typical/inc/os;../../src/BSP;../../src/FWlib;../../src/StartUP/ing918;../../src/Tools;../../src</IncludePath>
+              <IncludePath>.\libsbc\include;../../bundles/typical/inc;../../bundles/typical/inc/os;../../src/BSP;../../src/FWlib;../../src/StartUP/ing916;../../src/Tools;../../src</IncludePath>
             </VariousControls>
           </Cads>
           <Aads>
@@ -466,14 +466,14 @@
           <GroupName>StartUP</GroupName>
           <Files>
             <File>
-              <FileName>startup_ing91800.s</FileName>
+              <FileName>startup_ing91600.s</FileName>
               <FileType>2</FileType>
-              <FilePath>.\src\startup_ing91800.s</FilePath>
+              <FilePath>.\src\startup_ing91600.s</FilePath>
             </File>
             <File>
               <FileName>symdefs.m</FileName>
               <FileType>3</FileType>
-              <FilePath>..\..\bundles\typical\ING9188xx\symdefs.m</FilePath>
+              <FilePath>..\..\bundles\typical\ING9168xx\symdefs.m</FilePath>
             </File>
           </Files>
         </Group>
